[frontend] Extract graph controls from detail graph
The controls don't need to be tied to the graph itself. So we extract
them, reusing the component created for the comparison graph.
